refactor(deploy): improves structure
