+++ .eslintrc
@@ -14,7 +14,7 @@
         "paths": ["./", "src"]
       },
       "typescript": {
-        "project": "./src/tsconfig.json"
+        "project": "./tsconfig.server.json"
       }
     }
   }
