+++ .eslintrc.predev
@@ -1,7 +1,7 @@
 {
   "parser": "@typescript-eslint/parser",
   "parserOptions": {
-    "project": "./src/tsconfig.json"
+    "project": "./tsconfig.server.json"
   },
   "extends": [
     "@budproj/eslint-config-intl"
