--- .gitattributes
@@ -1,1 +1,0 @@
-/.yarn/** linguist-vendored
