+++ .gitignore
@@ -1,8 +1,6 @@
 # NextJS
 # --------------------------------------------------------------------------------------------------
 .next
-*/.next
-*/**/.next
 
 # Logs
 # --------------------------------------------------------------------------------------------------
@@ -20,3 +18,8 @@ compiled-lang
 # Environment
 # --------------------------------------------------------------------------------------------------
 .env
+
+# Deployment
+# --------------------------------------------------------------------------------------------------
+.waypoint
+dist
