+++ jest.config.ts
@@ -1,6 +1,7 @@
 export default {
   preset: 'ts-jest',
   setupFilesAfterEnv: ['./lib/jest/setup.ts'],
+  testPathIgnorePatterns: ['node_modules', 'dist'],
   moduleNameMapper: {
     '^src/(.*)$': '<rootDir>/src/$1',
     '^lib/(.*)$': '<rootDir>/lib/$1',
@@ -10,6 +11,7 @@ export default {
   },
   globals: {
     'ts-jest': {
+      tsconfig: 'tsconfig.server.json',
       babelConfig: {
         presets: ['@babel/preset-react'],
       },
