+++ src/components/Pages/KeyResults/index.ts
@@ -1,0 +1,3 @@
+import KeyResultsPage from './key-results'
+
+export default KeyResultsPage
