+++ src/components/Pages/KeyResults/key-results.spec.tsx
@@ -7,7 +7,9 @@ import sinon from 'sinon'
 
 import { keyResultOpenDrawer } from 'src/state/recoil/key-result/drawer'
 
-import KeyResultsPage, { messages } from '.'
+import KeyResultsPage from './key-results'
+
+import messages from './messages'
 
 describe('page control behaviors', () => {
   afterEach(() => sinon.restore())
