+++ src/components/Pages/KeyResults/key-results.tsx
@@ -1,0 +1,33 @@
+import React, { useEffect } from 'react'
+import { useIntl } from 'react-intl'
+import { useSetRecoilState } from 'recoil'
+
+import PageContent from 'src/components/Base/PageContent'
+import { KeyResultView } from 'src/components/KeyResult'
+import { KeyResultSingleDrawer } from 'src/components/KeyResult/Single'
+import { KeyResult } from 'src/components/KeyResult/types'
+import { keyResultOpenDrawer } from 'src/state/recoil/key-result/drawer'
+import { pageTitleAtom } from 'src/state/recoil/page'
+
+import messages from './messages'
+
+const KeyResultsPage = () => {
+  const intl = useIntl()
+  const setPageTitle = useSetRecoilState(pageTitleAtom)
+  const setOpenDrawer = useSetRecoilState(keyResultOpenDrawer)
+
+  const handleLineClick = (id: KeyResult['id']) => setOpenDrawer(id)
+
+  useEffect((): void => {
+    setPageTitle(intl.formatMessage(messages.pageTitle))
+  }, [intl, setPageTitle])
+
+  return (
+    <PageContent>
+      <KeyResultView onLineClick={handleLineClick} />
+      <KeyResultSingleDrawer />
+    </PageContent>
+  )
+}
+
+export default KeyResultsPage
