+++ src/state/recoil/key-result/progress-update/current-progress.ts
@@ -4,7 +4,7 @@ import { DefaultValue, selectorFamily } from 'recoil'
 import { KeyResult, ProgressReport } from 'src/components/KeyResult/types'
 import { PREFIX } from 'src/state/recoil/intl/constants'
 import { buildPartialSelector } from 'src/state/recoil/key-result/selectors'
-import { RecoilSpecificationGetter, RecoilSpecificationSetter } from 'state/recoil/types'
+import { RecoilSpecificationGetter, RecoilSpecificationSetter } from 'src/state/recoil/types'
 
 const KEY = `${PREFIX}::CURRENT_PROGRESS`
