--- src/themes/types.ts
@@ -1,1 +1,0 @@
-export type MUIStyle = Styles<Theme, Record<string, unknown>, string>
