+++ tsconfig.json
@@ -1,11 +1,42 @@
 {
-  "extends": "./src/tsconfig.json",
+  "extends": "@tsconfig/recommended/tsconfig.json",
   "compilerOptions": {
-    "isolatedModules": false,
-    "noEmit": false,
-    "module": "commonjs",
-    "outDir": "dist",
-    "target": "es2017",
-    "jsx": "react"
-  }
+    "target": "esnext",
+    "module": "esnext",
+    "jsx": "preserve",
+    "lib": ["dom", "dom.iterable", "esnext", "ESNext.Intl", "ES2018.Intl", "es2017"],
+    "moduleResolution": "node",
+    "baseUrl": "./",
+    "paths": {
+      "lib/*": ["./lib/*"],
+      "src/*": ["./src/*"]
+    },
+    "strict": true,
+    "allowJs": true,
+    "noEmit": true,
+    "allowSyntheticDefaultImports": true,
+    "esModuleInterop": true,
+    "skipLibCheck": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "isolatedModules": true,
+    "removeComments": false,
+    "preserveConstEnums": true,
+    "sourceMap": true,
+    "forceConsistentCasingInFileNames": true,
+    "resolveJsonModule": true
+  },
+  "exclude": [".next", "next.config.js", "*.spec.*"],
+  "include": [
+    "next-env.d.ts",
+    "*.ts",
+    "*.tsx",
+    "*.d.ts",
+    "**/*.ts",
+    "**/*.tsx",
+    "**/*.d.ts",
+    "./lib/*.ts",
+    "./lib/**/*.ts",
+    "./jest.config.ts"
+  ]
 }
