+++ tsconfig.server.json
@@ -1,0 +1,11 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "isolatedModules": false,
+    "noEmit": false,
+    "module": "commonjs",
+    "outDir": "dist",
+    "target": "es2017",
+    "jsx": "react"
+  }
+}
